requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
celery==5.3.6
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-spectacular==0.26.5
eventlet==0.33.3
exceptiongroup==1.1.3
future==0.18.3
greenlet==3.0.1
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
telegram==0.0.1
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
